Add llvm/circt 'main' image
Adds a Dockerfile for an image consisting of an installation of
llvm/circt off the 'main' branch (whatever it pulls when the
Dockerfile runs). This provides a lean-ish (~500MB) install of
circt-opt, circt-translate, and firtool that can either be wrapped in
a shell script to fake an install of these tools or can be used as a
base image.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>