Install lsb-release and software-properties-common for llvm.sh
The install script for clang uses lsb-release to find the version of
ubuntu it is installing on, and is called out inside the script as a
dependency.  This was already being installed in the integration image,
but missing from the ci image.  Explicitly installing these dependencies
in both images leaves nothing to chance.