Install Z3 to integration test and CI build images (#25)
Z3 is a state-of-the-art SMT solver widely employed in the development of verification tools.
As of now, when an image would install `clang-tidy` or `llvm`, `z3-4.8.7` would be downloaded as well as a dependency. Because this is fairly old version, CI would report errors when building https://github.com/llvm/circt/pull/3991 which I've been trying to merge.
I have therefore introduced a script download and install the current latest version of Z3 available. 
Because the repo's `README` suggests to test the images locally, I also added logic to detect the host machine's architecture and download or compile an appropriate release of CMake and Z3; previously `circt-integration-test` would fail to build on my ARM machine.