[gRPC] Add a flag to deal with cmake 4
Add a flag to bump the required minimum version of cmake. As per usual,
cmake has a different interpretation of 'minimum' than I do. Since they
dropped support in version 4 for features supported in versions <3.5 if
a project requires a minimum version of 3.1, cmake interprets that as it
using features they dropped which is not necessarily the case.